Crate rusoto_athena

Amazon Athena is an interactive query service that lets you use standard SQL to analyze data directly in Amazon S3. You can point Athena at your data in Amazon S3 and run ad-hoc queries and get results in seconds. Athena is serverless, so there is no infrastructure to set up or manage. You pay only for the queries you run. Athena scales automatically—executing queries in parallel—so results are fast, even with large datasets and complex queries. For more information, see What is Amazon Athena in the Amazon Athena User Guide.

If you connect to Athena using the JDBC driver, use version 1.1.0 of the driver or later with the Amazon Athena API. Earlier version drivers do not support the API. For more information and to download the driver, see Accessing Amazon Athena with JDBC.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

If you're using the service, you're probably looking for AthenaClient and Athena.

Structs

AthenaClient
A client for the Amazon Athena API.

BatchGetNamedQueryInput

BatchGetNamedQueryOutput

BatchGetQueryExecutionInput

BatchGetQueryExecutionOutput

ColumnInfo
Information about the columns in a query execution result.

CreateNamedQueryInput

CreateNamedQueryOutput

CreateWorkGroupInput

CreateWorkGroupOutput

Datum
A piece of data (a field in the table).

DeleteNamedQueryInput

DeleteNamedQueryOutput

DeleteWorkGroupInput

DeleteWorkGroupOutput

EncryptionConfiguration
If query results are encrypted in Amazon S3, indicates the encryption option used (for example, SSE-KMS or CSE-KMS) and key information.

GetNamedQueryInput

GetNamedQueryOutput

GetQueryExecutionInput

GetQueryExecutionOutput

GetQueryResultsInput

GetQueryResultsOutput

GetWorkGroupInput

GetWorkGroupOutput

ListNamedQueriesInput

ListNamedQueriesOutput

ListQueryExecutionsInput

ListQueryExecutionsOutput

ListTagsForResourceInput

ListTagsForResourceOutput

ListWorkGroupsInput

ListWorkGroupsOutput

NamedQuery
A query, where QueryString is the list of SQL query statements that comprise the query.

QueryExecution
Information about a single instance of a query execution.

QueryExecutionContext
The database in which the query execution occurs.

QueryExecutionStatistics
The amount of data scanned during the query execution and the amount of time that it took to execute, and the type of statement that was run.

QueryExecutionStatus
The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

ResultConfiguration
The location in Amazon S3 where query results are stored and the encryption option, if any, used for query results. These are known as "client-side settings". If workgroup settings override client-side settings, then the query uses the location for the query results and the encryption configuration that are specified for the workgroup.

ResultConfigurationUpdates
The information about the updates in the query results, such as output location and encryption configuration for the query results.

ResultSet
The metadata and rows that comprise a query result set. The metadata describes the column structure and data types.

ResultSetMetadata
The metadata that describes the column structure and data types of a table of query results.

Row
The rows that comprise a query result table.

StartQueryExecutionInput

StartQueryExecutionOutput

StopQueryExecutionInput

StopQueryExecutionOutput

Tag
A tag that you can add to a resource. A tag is a label that you assign to an AWS Athena resource (a workgroup). Each tag consists of a key and an optional value, both of which you define. Tags enable you to categorize workgroups in Athena, for example, by purpose, owner, or environment. Use a consistent set of tag keys to make it easier to search and filter workgroups in your account. The maximum tag key length is 128 Unicode characters in UTF-8. The maximum tag value length is 256 Unicode characters in UTF-8. You can use letters and numbers representable in UTF-8, and the following characters: + - = . _ : / @. Tag keys and values are case-sensitive. Tag keys must be unique per resource.

TagResourceInput

TagResourceOutput

UnprocessedNamedQueryId
Information about a named query ID that could not be processed.

UnprocessedQueryExecutionId
Describes a query execution that failed to process.

UntagResourceInput

UntagResourceOutput

UpdateWorkGroupInput

UpdateWorkGroupOutput

WorkGroup
A workgroup, which contains a name, description, creation time, state, and other configuration, listed under WorkGroup$Configuration. Each workgroup enables you to isolate queries for you or your group of users from other queries in the same account, to configure the query results location and the encryption configuration (known as workgroup settings), to enable sending query metrics to Amazon CloudWatch, and to establish per-query data usage control limits for all queries in a workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

WorkGroupConfiguration
The configuration of the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption option, if any, used for query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup and whether workgroup settings override query settings, and the data usage limit for the amount of data scanned per query, if it is specified. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

WorkGroupConfigurationUpdates
The configuration information that will be updated for this workgroup, which includes the location in Amazon S3 where query results are stored, the encryption option, if any, used for query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup, whether the workgroup settings override the client-side settings, and the data usage limit for the amount of bytes scanned per query, if it is specified.

WorkGroupSummary
The summary information for the workgroup, which includes its name, state, description, and the date and time it was created.

Enums

BatchGetNamedQueryError
Errors returned by BatchGetNamedQuery

BatchGetQueryExecutionError
Errors returned by BatchGetQueryExecution

CreateNamedQueryError
Errors returned by CreateNamedQuery

CreateWorkGroupError
Errors returned by CreateWorkGroup

DeleteNamedQueryError
Errors returned by DeleteNamedQuery

DeleteWorkGroupError
Errors returned by DeleteWorkGroup

GetNamedQueryError
Errors returned by GetNamedQuery

GetQueryExecutionError
Errors returned by GetQueryExecution

GetQueryResultsError
Errors returned by GetQueryResults

GetWorkGroupError
Errors returned by GetWorkGroup

ListNamedQueriesError
Errors returned by ListNamedQueries

ListQueryExecutionsError
Errors returned by ListQueryExecutions

ListTagsForResourceError
Errors returned by ListTagsForResource

ListWorkGroupsError
Errors returned by ListWorkGroups

StartQueryExecutionError
Errors returned by StartQueryExecution

StopQueryExecutionError
Errors returned by StopQueryExecution

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateWorkGroupError
Errors returned by UpdateWorkGroup

Traits

Athena
Trait representing the capabilities of the Amazon Athena API. Amazon Athena clients implement this trait.