Struct rusoto_athena::ColumnInfo

pub struct ColumnInfo {
    pub case_sensitive: Option<bool>,
    pub catalog_name: Option<String>,
    pub label: Option<String>,
    pub name: String,
    pub nullable: Option<String>,
    pub precision: Option<i64>,
    pub scale: Option<i64>,
    pub schema_name: Option<String>,
    pub table_name: Option<String>,
    pub type_: String,
}

Information about the columns in a query execution result.

Fields
case_sensitive: Option<bool>

Indicates whether values in the column are case-sensitive.

catalog_name: Option<String>

The catalog to which the query results belong.

label: Option<String>

A column label.

name: String

The name of the column.

nullable: Option<String>

Indicates the column's nullable status.

precision: Option<i64>

For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

scale: Option<i64>

For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

schema_name: Option<String>

The schema name (database name) to which the query results belong.

table_name: Option<String>

The table name for the query results.

type_: String

The data type of the column.