Trait rusoto_athena::Athena

pub trait Athena {
    fn batch_get_named_query(
        &self, 
        input: BatchGetNamedQueryInput
    ) -> RusotoFuture<BatchGetNamedQueryOutput, BatchGetNamedQueryError>;

    fn batch_get_query_execution(
        &self, 
        input: BatchGetQueryExecutionInput
    ) -> RusotoFuture<BatchGetQueryExecutionOutput, BatchGetQueryExecutionError>;

    fn create_named_query(
        &self, 
        input: CreateNamedQueryInput
    ) -> RusotoFuture<CreateNamedQueryOutput, CreateNamedQueryError>;

    fn create_work_group(
        &self, 
        input: CreateWorkGroupInput
    ) -> RusotoFuture<CreateWorkGroupOutput, CreateWorkGroupError>;

    fn delete_named_query(
        &self, 
        input: DeleteNamedQueryInput
    ) -> RusotoFuture<DeleteNamedQueryOutput, DeleteNamedQueryError>;

    fn delete_work_group(
        &self, 
        input: DeleteWorkGroupInput
    ) -> RusotoFuture<DeleteWorkGroupOutput, DeleteWorkGroupError>;

    fn get_named_query(
        &self, 
        input: GetNamedQueryInput
    ) -> RusotoFuture<GetNamedQueryOutput, GetNamedQueryError>;

    fn get_query_execution(
        &self, 
        input: GetQueryExecutionInput
    ) -> RusotoFuture<GetQueryExecutionOutput, GetQueryExecutionError>;

    fn get_query_results(
        &self, 
        input: GetQueryResultsInput
    ) -> RusotoFuture<GetQueryResultsOutput, GetQueryResultsError>;

    fn get_work_group(
        &self, 
        input: GetWorkGroupInput
    ) -> RusotoFuture<GetWorkGroupOutput, GetWorkGroupError>;

    fn list_named_queries(
        &self, 
        input: ListNamedQueriesInput
    ) -> RusotoFuture<ListNamedQueriesOutput, ListNamedQueriesError>;

    fn list_query_executions(
        &self, 
        input: ListQueryExecutionsInput
    ) -> RusotoFuture<ListQueryExecutionsOutput, ListQueryExecutionsError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceInput
    ) -> RusotoFuture<ListTagsForResourceOutput, ListTagsForResourceError>;

    fn list_work_groups(
        &self, 
        input: ListWorkGroupsInput
    ) -> RusotoFuture<ListWorkGroupsOutput, ListWorkGroupsError>;

    fn start_query_execution(
        &self, 
        input: StartQueryExecutionInput
    ) -> RusotoFuture<StartQueryExecutionOutput, StartQueryExecutionError>;

    fn stop_query_execution(
        &self, 
        input: StopQueryExecutionInput
    ) -> RusotoFuture<StopQueryExecutionOutput, StopQueryExecutionError>;

    fn tag_resource(
        &self, 
        input: TagResourceInput
    ) -> RusotoFuture<TagResourceOutput, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceInput
    ) -> RusotoFuture<UntagResourceOutput, UntagResourceError>;

    fn update_work_group(
        &self, 
        input: UpdateWorkGroupInput
    ) -> RusotoFuture<UpdateWorkGroupOutput, UpdateWorkGroupError>;
}

Trait representing the capabilities of the Amazon Athena API. Amazon Athena clients implement this trait.

Required methods

fn batch_get_named_query(
&self,
input: BatchGetNamedQueryInput
) -> RusotoFuture<BatchGetNamedQueryOutput, BatchGetNamedQueryError>

Returns the details of a single named query or a list of up to 50 queries, which you provide as an array of query ID strings. Requires you to have access to the workgroup in which the queries were saved. Use ListNamedQueriesInput to get the list of named query IDs in the specified workgroup. If information could not be retrieved for a submitted query ID, information about the query ID submitted is listed under UnprocessedNamedQueryId. Named queries differ from executed queries. Use BatchGetQueryExecutionInput to get details about each unique query execution, and ListQueryExecutionsInput to get a list of query execution IDs.

fn batch_get_query_execution(
&self,
input: BatchGetQueryExecutionInput
) -> RusotoFuture<BatchGetQueryExecutionOutput, BatchGetQueryExecutionError>

Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings. Requires you to have access to the workgroup in which the queries ran. To get a list of query execution IDs, use ListQueryExecutionsInput$WorkGroup. Query executions differ from named (saved) queries. Use BatchGetNamedQueryInput to get details about named queries.

fn create_named_query(
&self,
input: CreateNamedQueryInput
) -> RusotoFuture<CreateNamedQueryOutput, CreateNamedQueryError>

Creates a named query in the specified workgroup. Requires that you have access to the workgroup.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn create_work_group(
&self,
input: CreateWorkGroupInput
) -> RusotoFuture<CreateWorkGroupOutput, CreateWorkGroupError>

Creates a workgroup with the specified name.

fn delete_named_query(
&self,
input: DeleteNamedQueryInput
) -> RusotoFuture<DeleteNamedQueryOutput, DeleteNamedQueryError>

Deletes the named query if you have access to the workgroup in which the query was saved.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn delete_work_group(
&self,
input: DeleteWorkGroupInput
) -> RusotoFuture<DeleteWorkGroupOutput, DeleteWorkGroupError>

Deletes the workgroup with the specified name. The primary workgroup cannot be deleted.

fn get_named_query(
&self,
input: GetNamedQueryInput
) -> RusotoFuture<GetNamedQueryOutput, GetNamedQueryError>

Returns information about a single query. Requires that you have access to the workgroup in which the query was saved.

fn get_query_execution(
&self,
input: GetQueryExecutionInput
) -> RusotoFuture<GetQueryExecutionOutput, GetQueryExecutionError>

Returns information about a single execution of a query if you have access to the workgroup in which the query ran. Each time a query executes, information about the query execution is saved with a unique ID.

fn get_query_results(
&self,
input: GetQueryResultsInput
) -> RusotoFuture<GetQueryResultsOutput, GetQueryResultsError>

Returns the results of a single query execution specified by QueryExecutionId if you have access to the workgroup in which the query ran. This request does not execute the query but returns results. Use StartQueryExecution to run a query.

fn get_work_group(
&self,
input: GetWorkGroupInput
) -> RusotoFuture<GetWorkGroupOutput, GetWorkGroupError>

Returns information about the workgroup with the specified name.

fn list_named_queries(
&self,
input: ListNamedQueriesInput
) -> RusotoFuture<ListNamedQueriesOutput, ListNamedQueriesError>

Provides a list of available query IDs only for queries saved in the specified workgroup. Requires that you have access to the workgroup.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn list_query_executions(
&self,
input: ListQueryExecutionsInput
) -> RusotoFuture<ListQueryExecutionsOutput, ListQueryExecutionsError>

Provides a list of available query execution IDs for the queries in the specified workgroup. Requires you to have access to the workgroup in which the queries ran.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceInput
) -> RusotoFuture<ListTagsForResourceOutput, ListTagsForResourceError>

Lists the tags associated with this workgroup.

fn list_work_groups(
&self,
input: ListWorkGroupsInput
) -> RusotoFuture<ListWorkGroupsOutput, ListWorkGroupsError>

Lists available workgroups for the account.

fn start_query_execution(
&self,
input: StartQueryExecutionInput
) -> RusotoFuture<StartQueryExecutionOutput, StartQueryExecutionError>

Runs the SQL query statements contained in the Query. Requires you to have access to the workgroup in which the query ran.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn stop_query_execution(
&self,
input: StopQueryExecutionInput
) -> RusotoFuture<StopQueryExecutionOutput, StopQueryExecutionError>

Stops a query execution. Requires you to have access to the workgroup in which the query ran.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn tag_resource(
&self,
input: TagResourceInput
) -> RusotoFuture<TagResourceOutput, TagResourceError>

Adds one or more tags to the resource, such as a workgroup. A tag is a label that you assign to an AWS Athena resource (a workgroup). Each tag consists of a key and an optional value, both of which you define. Tags enable you to categorize resources (workgroups) in Athena, for example, by purpose, owner, or environment. Use a consistent set of tag keys to make it easier to search and filter workgroups in your account. For best practices, see AWS Tagging Strategies. The key length is from 1 (minimum) to 128 (maximum) Unicode characters in UTF-8. The tag value length is from 0 (minimum) to 256 (maximum) Unicode characters in UTF-8. You can use letters and numbers representable in UTF-8, and the following characters: + - = . _ : / @. Tag keys and values are case-sensitive. Tag keys must be unique per resource. If you specify more than one, separate them by commas.

fn untag_resource(
&self,
input: UntagResourceInput
) -> RusotoFuture<UntagResourceOutput, UntagResourceError>

Removes one or more tags from the workgroup resource. Takes as an input a list of TagKey Strings separated by commas, and removes their tags at the same time.

fn update_work_group(
&self,
input: UpdateWorkGroupInput
) -> RusotoFuture<UpdateWorkGroupOutput, UpdateWorkGroupError>

Updates the workgroup with the specified name. The workgroup's name cannot be changed.

Loading content...