Crate rusoto_athena

Amazon Athena is an interactive query service that lets you use standard SQL to analyze data directly in Amazon S3. You can point Athena at your data in Amazon S3 and run ad-hoc queries and get results in seconds. Athena is serverless, so there is no infrastructure to set up or manage. You pay only for the queries you run. Athena scales automatically—executing queries in parallel—so results are fast, even with large datasets and complex queries. For more information, see What is Amazon Athena in the Amazon Athena User Guide.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

If you're using the service, you're probably looking for AthenaClient and Athena.

Structs

AthenaClient
A client for the Amazon Athena API.

BatchGetNamedQueryInput

BatchGetNamedQueryOutput

BatchGetQueryExecutionInput

BatchGetQueryExecutionOutput

ColumnInfo
Information about the columns in a query execution result.

CreateNamedQueryInput

CreateNamedQueryOutput

Datum
A piece of data (a field in the table).

DeleteNamedQueryInput

DeleteNamedQueryOutput

EncryptionConfiguration
If query results are encrypted in Amazon S3, indicates the Amazon S3 encryption option used.

GetNamedQueryInput

GetNamedQueryOutput

GetQueryExecutionInput

GetQueryExecutionOutput

GetQueryResultsInput

GetQueryResultsOutput

ListNamedQueriesInput

ListNamedQueriesOutput

ListQueryExecutionsInput

ListQueryExecutionsOutput

NamedQuery
A query, where QueryString is the SQL query statements that comprise the query.

QueryExecution
Information about a single instance of a query execution.

QueryExecutionContext
The database in which the query execution occurs.

QueryExecutionStatistics
The amount of data scanned during the query execution and the amount of time that it took to execute.

QueryExecutionStatus
The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

ResultConfiguration
The location in Amazon S3 where query results are stored and the encryption option, if any, used for query results.

ResultSet
The metadata and rows that comprise a query result set. The metadata describes the column structure and data types.

ResultSetMetadata
The metadata that describes the column structure and data types of a table of query results.

Row
The rows that comprise a query result table.

StartQueryExecutionInput

StartQueryExecutionOutput

StopQueryExecutionInput

StopQueryExecutionOutput

UnprocessedNamedQueryId
Information about a named query ID that could not be processed.

UnprocessedQueryExecutionId
Describes a query execution that failed to process.

Enums

BatchGetNamedQueryError
Errors returned by BatchGetNamedQuery

BatchGetQueryExecutionError
Errors returned by BatchGetQueryExecution

CreateNamedQueryError
Errors returned by CreateNamedQuery

DeleteNamedQueryError
Errors returned by DeleteNamedQuery

GetNamedQueryError
Errors returned by GetNamedQuery

GetQueryExecutionError
Errors returned by GetQueryExecution

GetQueryResultsError
Errors returned by GetQueryResults

ListNamedQueriesError
Errors returned by ListNamedQueries

ListQueryExecutionsError
Errors returned by ListQueryExecutions

StartQueryExecutionError
Errors returned by StartQueryExecution

StopQueryExecutionError
Errors returned by StopQueryExecution

Traits

Athena
Trait representing the capabilities of the Amazon Athena API. Amazon Athena clients implement this trait.