Struct rusoto_athena::AthenaClient

pub struct AthenaClient { /* fields omitted */ }

A client for the Amazon Athena API.

Methods

impl AthenaClient

pub fn new(region: Region) -> AthenaClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AthenaClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,