Trait rusoto_athena::Athena

pub trait Athena {
    fn batch_get_named_query(
        &self, 
        input: BatchGetNamedQueryInput
    ) -> RusotoFuture<BatchGetNamedQueryOutput, BatchGetNamedQueryError>;

    fn batch_get_query_execution(
        &self, 
        input: BatchGetQueryExecutionInput
    ) -> RusotoFuture<BatchGetQueryExecutionOutput, BatchGetQueryExecutionError>;

    fn create_named_query(
        &self, 
        input: CreateNamedQueryInput
    ) -> RusotoFuture<CreateNamedQueryOutput, CreateNamedQueryError>;

    fn delete_named_query(
        &self, 
        input: DeleteNamedQueryInput
    ) -> RusotoFuture<DeleteNamedQueryOutput, DeleteNamedQueryError>;

    fn get_named_query(
        &self, 
        input: GetNamedQueryInput
    ) -> RusotoFuture<GetNamedQueryOutput, GetNamedQueryError>;

    fn get_query_execution(
        &self, 
        input: GetQueryExecutionInput
    ) -> RusotoFuture<GetQueryExecutionOutput, GetQueryExecutionError>;

    fn get_query_results(
        &self, 
        input: GetQueryResultsInput
    ) -> RusotoFuture<GetQueryResultsOutput, GetQueryResultsError>;

    fn list_named_queries(
        &self, 
        input: ListNamedQueriesInput
    ) -> RusotoFuture<ListNamedQueriesOutput, ListNamedQueriesError>;

    fn list_query_executions(
        &self, 
        input: ListQueryExecutionsInput
    ) -> RusotoFuture<ListQueryExecutionsOutput, ListQueryExecutionsError>;

    fn start_query_execution(
        &self, 
        input: StartQueryExecutionInput
    ) -> RusotoFuture<StartQueryExecutionOutput, StartQueryExecutionError>;

    fn stop_query_execution(
        &self, 
        input: StopQueryExecutionInput
    ) -> RusotoFuture<StopQueryExecutionOutput, StopQueryExecutionError>;
}

Trait representing the capabilities of the Amazon Athena API. Amazon Athena clients implement this trait.

Required Methods

fn batch_get_named_query(
&self,
input: BatchGetNamedQueryInput
) -> RusotoFuture<BatchGetNamedQueryOutput, BatchGetNamedQueryError>

Returns the details of a single named query or a list of up to 50 queries, which you provide as an array of query ID strings. Use ListNamedQueries to get the list of named query IDs. If information could not be retrieved for a submitted query ID, information about the query ID submitted is listed under UnprocessedNamedQueryId. Named queries are different from executed queries. Use BatchGetQueryExecution to get details about each unique query execution, and ListQueryExecutions to get a list of query execution IDs.

fn batch_get_query_execution(
&self,
input: BatchGetQueryExecutionInput
) -> RusotoFuture<BatchGetQueryExecutionOutput, BatchGetQueryExecutionError>

Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings. To get a list of query execution IDs, use ListQueryExecutions. Query executions are different from named (saved) queries. Use BatchGetNamedQuery to get details about named queries.

fn create_named_query(
&self,
input: CreateNamedQueryInput
) -> RusotoFuture<CreateNamedQueryOutput, CreateNamedQueryError>

Creates a named query.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn delete_named_query(
&self,
input: DeleteNamedQueryInput
) -> RusotoFuture<DeleteNamedQueryOutput, DeleteNamedQueryError>

Deletes a named query.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn get_named_query(
&self,
input: GetNamedQueryInput
) -> RusotoFuture<GetNamedQueryOutput, GetNamedQueryError>

Returns information about a single query.

fn get_query_execution(
&self,
input: GetQueryExecutionInput
) -> RusotoFuture<GetQueryExecutionOutput, GetQueryExecutionError>

Returns information about a single execution of a query. Each time a query executes, information about the query execution is saved with a unique ID.

fn get_query_results(
&self,
input: GetQueryResultsInput
) -> RusotoFuture<GetQueryResultsOutput, GetQueryResultsError>

Returns the results of a single query execution specified by QueryExecutionId. This request does not execute the query but returns results. Use StartQueryExecution to run a query.

fn list_named_queries(
&self,
input: ListNamedQueriesInput
) -> RusotoFuture<ListNamedQueriesOutput, ListNamedQueriesError>

Provides a list of all available query IDs.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn list_query_executions(
&self,
input: ListQueryExecutionsInput
) -> RusotoFuture<ListQueryExecutionsOutput, ListQueryExecutionsError>

Provides a list of all available query execution IDs.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn start_query_execution(
&self,
input: StartQueryExecutionInput
) -> RusotoFuture<StartQueryExecutionOutput, StartQueryExecutionError>

Runs (executes) the SQL query statements contained in the Query string.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

fn stop_query_execution(
&self,
input: StopQueryExecutionInput
) -> RusotoFuture<StopQueryExecutionOutput, StopQueryExecutionError>

Stops a query execution.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.